Update sendmail and fixed bugs

+++ app/Http/Controllers/OrdersController.php
@@ -59,16 +59,26 @@ class OrdersController extends Controller
     }
 
     public function getdel($id)
-    {       
+    {
     	$order = Orders::where('id',$id)->first();
-    	if ($order->status ==1) {
+        if ($order == null) {
+            return redirect('admin/donhang');
+        }
+
+    	if ($order->status == 1) {
     		return redirect()->back()
-    		->with(['flash_level' => 'result_msg', 'flash_massage' => 'Không thể hủy đơn hàng số: '. $id . ' vì đã được xác nhận!']);
+    		->with([
+                'flash_level' => 'result_msg',
+                'flash_massage' => 'Không thể hủy đơn hàng số: '. $id . ' vì đã được xác nhận!'
+            ]);
     	} else {
     		$order = Orders::find($id);
         	$order->delete();
         	return redirect('admin/donhang')
-         	->with(['flash_level'=>'result_msg','flash_massage'=>'Đã hủy bỏ đơn hàng số:  '.$id.'!']);
+         	->with([
+                'flash_level'=>'result_msg',
+                'flash_massage'=>'Đã hủy bỏ đơn hàng số:  '.$id.'!'
+            ]);
      	}
     }
 }
